requirements-dev.txt
coverage==4.5.4
ipdb==0.12.2
ipython==7.8
pytest-django>=3.5.1
pytest==5.1.2
pytest-cov==2.7.1
flake8==3.7.8
django-extensions==2.2.1
tox

-e .
-e demo
